Add `.gitattributes` file
Adds a `.gitattributes` file with `export-ignore` and `diff` rules [suitable for composer projects](https://php.watch/articles/composer-gitattributes)

Creating a project with `composer create-project` still clones the entire repository, so all files get downloaded. However, when the package is published, consumers of that package will not get the build files if they download it with `prefer-dist`, which is the default.
This can reduce the download size of packages by a significant percentage.